Applied Refactoring
