+++ core/pom.xml
@@ -130,6 +130,12 @@
 			<version>3.6.1</version>
 			<scope>compile</scope>
 		</dependency>
+		<dependency>
+			<groupId>it.unimi.dsi</groupId>
+			<artifactId>fastutil</artifactId>
+			<version>7.0.13</version>
+			<scope>compile</scope>
+		</dependency>
 
 		<!-- test scope -->
 		<dependency>
