--- core/src/test/resources/hivemall/classifier/news20-tiny.binary
@@ -1,2 +1,0 @@
--1 75:0.097129 97:0.097129 244:0.097129 1207:0.097129 2013:0.097129 2317:0.097129 2781:0.097129 3014:0.097129 3517:0.097129 3539:0.097129 3772:0.097129 3813:0.097129 3831:0.097129 3838:0.097129 4055:0.097129 4118:0.097129 4121:0.097129 4203:0.097129 4664:0.097129 5252:0.097129 6127:0.097129 6396:0.097129 6667:0.097129 6862:0.097129 7643:0.097129 8302:0.097129 8851:0.097129 9049:0.097129 9059:0.097129 9119:0.097129 9135:0.097129 10263:0.097129 11409:0.097129 12458:0.097129 12459:0.097129 12460:0.097129 12461:0.097129 12462:0.097129 12463:0.097129 12464:0.097129 12465:0.097129 12466:0.097129 12467:0.097129 12468:0.097129 12469:0.097129 12470:0.097129 12471:0.097129 12472:0.097129 12473:0.097129 12474:0.097129 12475:0.097129 12476:0.097129 12477:0.097129 12478:0.097129 12479:0.097129 12480:0.097129 12481:0.097129 12482:0.097129 12483:0.097129 12484:0.097129 12485:0.097129 12486:0.097129 12487:0.097129 12488:0.097129 12489:0.097129 12490:0.097129 12491:0.097129 12492:0.097129 12493:0.097129 12494:0.097129 12495:0.097129 12496:0.097129 12497:0.097129 12498:0.097129 12499:0.097129 12500:0.097129 12501:0.097129 12502:0.097129 12503:0.097129 12504:0.097129 12505:0.097129 12506:0.097129 12507:0.097129 12508:0.097129 12509:0.097129 12510:0.097129 12511:0.097129 12512:0.097129 12513:0.097129 12514:0.097129 12515:0.097129 12516:0.097129 12517:0.097129 12518:0.097129 12519:0.097129 12520:0.097129 12521:0.097129 12522:0.097129 12523:0.097129 12524:0.097129 12525:0.097129 12526:0.097129 12527:0.097129 12528:0.097129 12529:0.097129 12530:0.097129
-+1 54:0.158114 114:0.158114 480:0.158114 1977:0.158114 5818:0.158114 9072:0.158114 19599:0.158114 28803:0.158114 33107:0.158114 181139:0.158114 181168:0.158114 204449:0.158114 222034:0.158114 222396:0.158114 225771:0.158114 226795:0.158114 232824:0.158114 252132:0.158114 256786:0.158114 256787:0.158114 256788:0.158114 256789:0.158114 256790:0.158114 256791:0.158114 256792:0.158114 256793:0.158114 256794:0.158114 256795:0.158114 256796:0.158114 256797:0.158114 256798:0.158114 256799:0.158114 256800:0.158114 256801:0.158114 256802:0.158114 256803:0.158114 256804:0.158114 256805:0.158114 256806:0.158114 256807:0.158114
